+++ node/standalone/Cargo.toml
@@ -41,14 +41,14 @@ sp-runtime = "3.0.0"
 sp-transaction-pool = "3.0.0"
 structopt = "0.3.21"
 substrate-frame-rpc-system = "3.0.0"
-vln-runtime = { path = '../../runtime/standalone', version = '2.0.0' }
+vln-runtime = { package = "vln-parachain-runtime", path = '../../runtime/', features = ['standalone'], version = '2.0.0' }
 
 [build-dependencies]
 substrate-build-script-utils = "3.0.0"
 
 [features]
 default = []
-native-runtime-benchmarks = ['runtime-benchmarks', 'vln-runtime/native-runtime-benchmarks']
+native-runtime-benchmarks = ['runtime-benchmarks']
 runtime-benchmarks = [
     'frame-benchmarking-cli',
     'frame-benchmarking',
