--- runtime/standalone/build.rs
@@ -1,11 +1,0 @@
-//! Wasm builder
-
-use substrate_wasm_builder::WasmBuilder;
-
-fn main() {
-    WasmBuilder::new()
-        .with_current_project()
-        .export_heap_base()
-        .import_memory()
-        .build()
-}
